requirements.txt
crewai==0.100.1
crewai_tools==0.33.0
fastapi==0.109.2
groq==0.17.0
pydantic==2.10.6
python-dotenv==1.0.1
langchain-community==0.3.16
duckduckgo_search==7.3.0
uvicorn==0.27.1
